Update coinbase.py
I've included two different spreads. For some reason there are two different ways to look up the exchange rate of a currency, and they seem to update at different rates. The exchange rate from ETH to USD is also identical to the rate from (USD to ETH)/1. However, the descriptions for buying and selling ETH from USD say that there's an included 1% fee, so they come up with different results. As a result I added a 1% fee to the exchange rate method.